backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
SQLAlchemy
requests
beautifulsoup4
wechatpy
python-json-logger
tqdm
whisper
jieba
summa